+++ packages/agility_redux_bloc/example/lib/counter_bloc/counter_bloc.dart
@@ -1,6 +1,7 @@
 import 'package:agility_redux_bloc/agility_redux_bloc.dart';
 import 'package:flutter/material.dart';
 
+import 'actionsheet/action_sheet_popup_box.dart';
 import 'counter_page/counter_page.dart';
 import 'counter_redux_bloc.dart';
 
@@ -22,6 +23,8 @@ class CounterBloc extends Bloc {
 
   @override
   Map<String, BlocPopupBoxBuilder> initialPopupBoxList() {
-    return {};
+    return {
+      'action_sheet': (RouteSettings settings) => ActionSheetPopupBox(settings: settings),
+    };
   }
 }