+++ packages/agility_redux_bloc/example/lib/counter_bloc/counter_page/counter_widget.dart
@@ -1,4 +1,5 @@
 import 'package:agility_redux_bloc/agility_redux_bloc.dart';
+import 'package:agility_redux_widget/agility_redux_widget.dart';
 import 'package:flutter/material.dart';
 
 import '../counter_redux_bloc.dart';
@@ -12,28 +13,49 @@ class CounterWidget extends StatelessWidget {
         title: Text('Counter'),
       ),
       body: Center(
-        child: ViewModelSubscriber<CounterViewModel>(
-            converter: (state) => CounterViewModel.fromState(state),
-            builder: (context, dispatcher, viewModel) {
-              return Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Text(
-                    'You have pushed the button this many times:',
-                  ),
-                  Text(
-                    '${viewModel.number}',
-                    style: Theme.of(context).textTheme.headline4,
-                  ),
-                  FlatButton(
-                    onPressed: () {
-                      GlobalNavigator().navigatorEntry('page').push('counter/counter_page', holdBlocNames: ['counter']);
-                    },
-                    child: Text('push new page'),
-                  ),
-                ],
-              );
-            }),
+        child: ActionSubscriber<CounterAddAction>(
+          onReceivedAction: (action) {
+            print('========== action $action');
+          },
+          child: Builder(builder: (context) {
+            print('---&&&&--- build');
+            return ViewModelSubscriber<CounterViewModel>(
+                converter: (state) => CounterViewModel.fromState(state),
+                builder: (context, dispatcher, viewModel) {
+                  print('---+++---- build');
+                  return Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      // ActionSubscriber<CounterAddAction>(
+                      //   onReceivedAction: (action) {
+                      //     print('========== action $action');
+                      //   },
+                      //   child: Builder(
+                      //     builder: (context) {
+                      //       print('------- build');
+                      //       return Container();
+                      //     },
+                      //   ),
+                      // ),
+                      Text(
+                        'You have pushed the button this many times:',
+                      ),
+                      Text(
+                        '${viewModel.number}',
+                        style: Theme.of(context).textTheme.headline4,
+                      ),
+                      FlatButton(
+                        onPressed: () {
+                          // GlobalNavigator().navigatorEntry('page').push('counter/counter_page', holdBlocNames: ['counter']);
+                          GlobalNavigator().navigatorEntry('popupBox').push('counter/action_sheet');
+                        },
+                        child: Text('push new page'),
+                      ),
+                    ],
+                  );
+                });
+          }),
+        ),
       ),
       floatingActionButton: DispatchSubscriber(builder: (context, dispatcher) {
         return FloatingActionButton(